requirements.txt
Pillow==9.3.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
pycocotools==2.0.6
ultralytics==8.0.50